refactor(react-suspense-cache): make suspense explicit with use
