+++ packages/react-suspense-cache/src/index.ts
@@ -2,6 +2,5 @@ export * from "./constants.js";
 export * from "./createCache.js";
 export * from "./createExternallyManagedCache.js";
 export * from "./deferred.js";
-export * from "./hooks.js";
 export * from "./records.js";
 export * from "./types.js";
